reallyroutespline: squash a -Wsign-conversion warning
By construction, `inpn` can only ever be non-negative here. Though in the long
term a more sweeping change should be done to count items in `size_t` variables
pervasively.

+++ lib/pathplan/route.c
@@ -105,7 +105,7 @@ static int reallyroutespline(Pedge_t * edges, int edgen,
     static int tnan;
 
     if (tnan < inpn) {
-	if (!(tnas = realloc(tnas, sizeof(tna_t) * inpn)))
+	if (!(tnas = realloc(tnas, sizeof(tna_t) * (size_t)inpn)))
 	    return -1;
 	tnan = inpn;
     }
